// src/types.ts
// src/types.ts
export type CityData = {
    Name: string;
    Latitude: number;
    Longitude: number;
    URL: string;
    GreenSpacePercentage: number;
    GreenSpacePercentage_Trees: number;
    GreenSpacePercentage_Grass: number;
    VegetationHealth: number;
    GreenSpaceDistribution: number;
  };